feat: create insane data-structure for day7
To be honest, I think the heart of the problem is yet another stack
problem. I think I could accomplish the same thing just using a handful
of running totals.

This was good practice and exercise, though. Hopefully I can figure out
how to use this data-structure to produce the desired output...