perf: move tail `Knot` pos tracker into `Rope`
For the sake of this exercise it's overkill for all `Knots` to keep
track of all of their visited locations. It makes more sense for the
`Rope` to keep track of that information (especially due to only needing
to track the last `Knot`'s positions).

I also extracted some methods and reduced repetition where possible.
I would have liked to have done this in another commit, but the changes
were too closely coupled. I prefer to only commit working code, and
breaking apart the commit may have violated that.