feat: solve day 12 part 1
I was misreading the instructions which led to an off-by-one error.
I've been beating my head against the wall trying to figure out what was
wrong with my Djikstra's, and it was a basic off-by-one.

Thank you Charles Taylor for pointing this out to me.
I probably would have re-written the entire thing again if you hadn't.